Add more controls over status update and comments
Added "On failure" and "Never" options to "Report status on" and
enhanced "Use Comments" from being a checkbox to being selection
property that allows options like "Always", "With status", and "Never".
This combination of controls allows flexiblity about when and how to
update the status of a PR in github when the build is part of a chain
and/or when multiple independent builds need to report status as
comments, but where their result should not indicate overall status of
the PR (i.e. information only).

Also, ensured that migration of the comments checkbox works correctly.
If it was checked previously, that becomes "with status" to match
previous behavior.  If it was unchecked previously, it becomes "never",
again to maintain prior behavior.